Add to paragraph
I noted and added to the paragraph text that I have not found
anything suitable that would limit the body text to seventy
characters. I need to return to this issue. Perhaps, if laer
I refeactor git with the commit template. The original
templalte usees hashtags to hide the instructive section of
template